requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.15
decorator==5.2.1
dnspython==2.7.0
executing==2.2.0
flake8==7.3.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pymongo==4.10.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==27.0.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
